myproject/requirements.txt
asgiref==3.7.2
cmdstanpy==1.1.0
colorama==0.4.6
contourpy==1.1.0
convertdate==2.4.0
cycler==0.11.0
Django==4.2.2
djangorestframework==3.14.0
ephem==4.1.4
fonttools==4.40.0
holidays==0.26
importlib-resources==5.12.0
kiwisolver==1.4.4
LunarCalendar==0.0.9
matplotlib==3.7.1
numpy==1.25.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
prophet==1.1.4
PyMeeus==0.5.12
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
zipp==3.15.0
